Split old osm test file to dedicated topic files
